% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_words.R
\name{match_words}
\alias{match_words}
\title{Match words with dictionary}
\usage{
match_words(word_vector, dictionary)
}
\arguments{
\item{words}{Vector of preprocessed words.}

\item{patterns}{Vector of patterns, may contain "*" as wildcard.}
}
\value{
Number of matched words.
}
\description{
Counts how many words from the text match a given set of patterns
(supports prefix or suffix wildcards like "sad*" or "*ing").
}
