% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_text.R
\name{preprocess_text}
\alias{preprocess_text}
\title{Preprocess text file}
\usage{
preprocess_text(filepath)
}
\arguments{
\item{path}{Path to a text file.}
}
\value{
A character vector of words.
}
\description{
Loads a .txt file, converts to lowercase, removes punctuation,
and splits into individual words.
}
