% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_sentiment.R
\name{summarize_sentiment}
\alias{summarize_sentiment}
\title{Summarize sentiment counts}
\usage{
summarize_sentiment(word_vector, positive_dict, negative_dict)
}
\arguments{
\item{words}{Vector of text words.}

\item{positive}{Vector of positive patterns.}

\item{negative}{Vector of negative patterns.}
}
\value{
A list with positive, negative, and ratio.
}
\description{
Computes the number of positive and negative matches and returns a summary.
}
